Replaced bb-sync-dir-one-way function by flag of bb-sync-dir

+++ source/30_sync.sh
@@ -24,12 +24,31 @@ bb-sync-file() {
     fi
 }
 
+bb-sync-dir() {
+    local TWO_WAY=true
+    local -a ARGS=()
+    local -i TEST=3    # Test first three arguments
+
+    while (( $# && TEST-- ))
+    do
+        case "$1" in
+            -o|--one-way) TWO_WAY=false ;;
+            -t|--two-way) TWO_WAY=true  ;;
+            *) ARGS[${#ARGS[@]}]="$1"   ;;
+        esac
+        shift
+    done
+    ARGS=( "${ARGS[@]}" "$@" )
+    echo "${ARGS[@]}"
+
+    bb-sync-dir-helper "$TWO_WAY" "${ARGS[@]}"
+}
+
 bb-sync-dir-helper() {
-    local ONE_WAY="$1"
-    shift
-    local DST_DIR="$( readlink -nm "$1" )"
-    local SRC_DIR="$( readlink -ne "$2" )"
-    shift 2
+    local TWO_WAY="$1"
+    local DST_DIR="$( readlink -nm "$2" )"
+    local SRC_DIR="$( readlink -ne "$3" )"
+    shift 3
 
     if [[ ! -d "$DST_DIR" ]]
     then
@@ -49,33 +68,28 @@ bb-sync-dir-helper() {
             bb-sync-file "$DST_DIR/$NAME" "$SRC_DIR/$NAME" "$@"
         elif [[ -d "$SRC_DIR/$NAME" ]]
         then
-            bb-sync-dir-helper "$ONE_WAY" "$DST_DIR/$NAME" "$SRC_DIR/$NAME" "$@"
+            bb-sync-dir-helper "$TWO_WAY" "$DST_DIR/$NAME" "$SRC_DIR/$NAME" "$@"
         fi
     done < <( ls -A )
-    cd "$DST_DIR"
-    while [ "$ONE_WAY" -eq 0 ] && read -r NAME
-    do
-        if [[ ! -e "$SRC_DIR/$NAME" ]]
-        then
-            local EVENT="bb-sync-file-removed"
-            if [[ -d "$DST_DIR/$NAME" ]]
+
+    if $TWO_WAY
+    then
+        cd "$DST_DIR"
+        while read -r NAME
+        do
+            if [[ ! -e "$SRC_DIR/$NAME" ]]
             then
-                EVENT="bb-sync-dir-removed"
+                local EVENT="bb-sync-file-removed"
+                if [[ -d "$DST_DIR/$NAME" ]]
+                then
+                    EVENT="bb-sync-dir-removed"
+                fi
+                rm -rf "$DST_DIR/$NAME"
+                bb-event-delay "$@"
+                bb-event-fire "$EVENT" "$DST_DIR/$NAME"
             fi
-            rm -rf "$DST_DIR/$NAME"
-            bb-event-delay "$@"
-            bb-event-fire "$EVENT" "$DST_DIR/$NAME"
-        fi
-    done < <( find . )
+        done < <( find . )
+    fi
 
     cd "$ORIGINAL_DIR"
 }
-
-bb-sync-dir() {
-    bb-sync-dir-helper 0 "$@"
-}
-
-bb-sync-dir-one-way() {
-    bb-sync-dir-helper 1 "$@"
-}
-
